fix: update cui url for newer versions
