+++ src/constants.js
@@ -1,0 +1,1 @@
+export const EMBED_URL = 'https://public-assets.typeform.com/confab/embed.js'
